scylla-gdb: print number of allocated objects as an integer instead of float
Message-Id: <20171010151835.GT23527@scylladb.com>

+++ scylla-gdb.py
@@ -221,7 +221,7 @@ class scylla_memory(gdb.Command):
             pages_in_use = int(sp['_pages_in_use'])
             memory = pages_in_use * page_size
             # use_count can be off if we used fallback spans rather than preferred spans
-            use_count = (memory / span_size) * int(span_size / object_size) - free_count
+            use_count = int(memory / span_size) * int(span_size / object_size) - free_count
             wasted = free_count * object_size
             wasted_percent = wasted * 100.0 / memory if memory else 0
             gdb.write('{objsize:5} {span_size:6} {use_count:10} {memory:12} {wasted_percent:5.1f}\n'
